Tizen 2.4 SDK Rev6 Release

+++ misc/fchflags.c
@@ -22,12 +22,12 @@
 
 /* Change the flags of the file FD refers to to FLAGS.  */
 
-int fchflags (int fd, int flags) __THROW;
+int fchflags (int fd, unsigned long int flags) __THROW;
 
 int
 fchflags (fd, flags)
      int fd;
-     int flags;
+     unsigned long int flags;
 {
   if (fd < 0)
     {
