+++ sysdeps/mach/hurd/fchflags.c
@@ -26,7 +26,7 @@
 
 /* XXX should be __fchflags? */
 int
-fchflags (int fd, int flags)
+fchflags (int fd, unsigned long int flags)
 {
   error_t err;
